+++ include/smart_json/smart_json.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
 #include "camel_case_transformer.hpp"
-#include "decode.inl"
-#include "encode.inl"
+#include "decode.hpp"
+#include "encode.hpp"
 #include "no_transformer.hpp"